refactor: move test cases to RoutingInfoInitializationMigrationTest
